[oh-tab-zhwi.5.1] Decouple shared settings message types
